move the auth_token parameter to the query instead of the body
after speaking with maropost tech support and them showing me how they
generate api requests, it seems that the supported way of using the api
is via having the auth_token as a url query parameter, instead of part
of the json payload.

testing it on our personal account seems to work properly. originally i
was having problems updating custom fields of contacts that are in the
dnm list, but after this change i was able to.